+++ app/src/main/java/me/mikecasper/musicvoice/MusicVoiceApplication.java
@@ -2,6 +2,10 @@ package me.mikecasper.musicvoice;
 
 import android.app.Application;
 
+import me.mikecasper.musicvoice.services.eventmanager.EventManagerProvider;
+import me.mikecasper.musicvoice.services.eventmanager.IEventManager;
+import me.mikecasper.musicvoice.services.musicplayer.events.DestroyPlayerEvent;
+
 public class MusicVoiceApplication extends Application {
 
     public static final LogLevel LOG_LEVEL = LogLevel.DEBUG;
@@ -15,5 +19,13 @@ public class MusicVoiceApplication extends Application {
     @Override
     public void onCreate() {
         super.onCreate();
+    }
+
+    @Override
+    public void onTerminate() {
+        IEventManager eventManager = EventManagerProvider.getInstance(this);
+        eventManager.postEvent(new DestroyPlayerEvent());
+
+        super.onTerminate();
     }
 }
