+++ app/src/main/java/me/mikecasper/musicvoice/services/musicplayer/MusicPlayer.java
@@ -1,5 +1,8 @@
 package me.mikecasper.musicvoice.services.musicplayer;
 
+import android.os.Handler;
+import android.os.SystemClock;
+
 import com.spotify.sdk.android.player.Config;
 import com.spotify.sdk.android.player.ConnectionStateCallback;
 import com.spotify.sdk.android.player.Player;
@@ -11,6 +14,10 @@ import com.squareup.otto.Subscribe;
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
+import java.util.concurrent.Executors;
+import java.util.concurrent.ScheduledExecutorService;
+import java.util.concurrent.ScheduledFuture;
+import java.util.concurrent.TimeUnit;
 
 import me.mikecasper.musicvoice.api.responses.TrackResponseItem;
 import me.mikecasper.musicvoice.api.services.LogInService;
@@ -18,6 +25,7 @@ import me.mikecasper.musicvoice.models.Track;
 import me.mikecasper.musicvoice.nowplaying.NowPlayingActivity;
 import me.mikecasper.musicvoice.services.eventmanager.IEventManager;
 import me.mikecasper.musicvoice.services.musicplayer.events.CreatePlayerEvent;
+import me.mikecasper.musicvoice.services.musicplayer.events.DestroyPlayerEvent;
 import me.mikecasper.musicvoice.services.musicplayer.events.GetPlayerStatusEvent;
 import me.mikecasper.musicvoice.services.musicplayer.events.LostPermissionEvent;
 import me.mikecasper.musicvoice.services.musicplayer.events.PlaySongEvent;
@@ -26,10 +34,12 @@ import me.mikecasper.musicvoice.services.musicplayer.events.SetPlaylistEvent;
 import me.mikecasper.musicvoice.services.musicplayer.events.SkipBackwardEvent;
 import me.mikecasper.musicvoice.services.musicplayer.events.SkipForwardEvent;
 import me.mikecasper.musicvoice.services.musicplayer.events.SongChangeEvent;
+import me.mikecasper.musicvoice.services.musicplayer.events.StopSeekbarUpdateEvent;
 import me.mikecasper.musicvoice.services.musicplayer.events.TogglePlaybackEvent;
 import me.mikecasper.musicvoice.services.musicplayer.events.ToggleRepeatEvent;
 import me.mikecasper.musicvoice.services.musicplayer.events.ToggleShuffleEvent;
 import me.mikecasper.musicvoice.services.musicplayer.events.UpdatePlayerStatusEvent;
+import me.mikecasper.musicvoice.services.musicplayer.events.UpdateSongTimeEvent;
 import me.mikecasper.musicvoice.util.Logger;
 
 public class MusicPlayer implements ConnectionStateCallback, PlayerNotificationCallback {
@@ -46,6 +56,23 @@ public class MusicPlayer implements ConnectionStateCallback, PlayerNotificationC
     private List<Track> mTracks;
     private List<Track> mOriginalTracks;
     private IEventManager mEventManager;
+
+    // Song time stuff
+    private static final long PROGRESS_UPDATE_INTERNAL = 1000;
+    private static final long PROGRESS_UPDATE_INITIAL_INTERVAL = 100;
+
+    private ScheduledFuture<?> mScheduledFuture;
+
+    private final ScheduledExecutorService mExecutorService = Executors.newSingleThreadScheduledExecutor();
+    private final Handler mHandler = new Handler();
+    private long mPreviousTime;
+    private int mPreviousSongTime;
+    private final Runnable mUpdateProgressTask = new Runnable() {
+        @Override
+        public void run() {
+            updateProgress();
+        }
+    };
 
     public MusicPlayer(IEventManager eventManager) {
         mEventManager = eventManager;
@@ -90,6 +117,11 @@ public class MusicPlayer implements ConnectionStateCallback, PlayerNotificationC
     public void onPlaySongEvent(PlaySongEvent event) {
         mPlayer.play(mTracks.get(0).getUri());
         mIsPlaying = true;
+
+        mPreviousSongTime = 0;
+
+        mPreviousTime = SystemClock.elapsedRealtime();
+        scheduleSeekBarUpdate();
     }
 
     @Subscribe
@@ -143,8 +175,11 @@ public class MusicPlayer implements ConnectionStateCallback, PlayerNotificationC
     public void onTogglePlayback(TogglePlaybackEvent event) {
         if (mIsPlaying) {
             mPlayer.pause();
+            stopSeekBarUpdate();
         } else {
             mPlayer.resume();
+            mPreviousTime = SystemClock.elapsedRealtime();
+            scheduleSeekBarUpdate();
         }
 
         mIsPlaying = !mIsPlaying;
@@ -153,6 +188,8 @@ public class MusicPlayer implements ConnectionStateCallback, PlayerNotificationC
     @Subscribe
     public void onSeekTo(SeekToEvent event) {
         mPlayer.seekToPosition(event.getPosition());
+        mPreviousSongTime = event.getPosition();
+        scheduleSeekBarUpdate();
     }
 
     @Subscribe
@@ -177,6 +214,11 @@ public class MusicPlayer implements ConnectionStateCallback, PlayerNotificationC
         playPreviousSong();
     }
 
+    @Subscribe
+    public void onStopSeekUpdates(StopSeekbarUpdateEvent event) {
+        stopSeekBarUpdate();
+    }
+
     private void playNextSong() {
         boolean shouldPlaySong = true;
 
@@ -199,6 +241,7 @@ public class MusicPlayer implements ConnectionStateCallback, PlayerNotificationC
         }
 
         mEventManager.postEvent(new SongChangeEvent(mTracks.get(mSongIndex), shouldPlaySong));
+        updateTask();
     }
 
     private void playPreviousSong() {
@@ -231,6 +274,18 @@ public class MusicPlayer implements ConnectionStateCallback, PlayerNotificationC
         }
 
         mEventManager.postEvent(new SongChangeEvent(mTracks.get(mSongIndex), shouldPlaySong));
+        updateTask();
+    }
+
+    private void updateTask() {
+        mPreviousSongTime = 0;
+
+        if (mIsPlaying) {
+            mPreviousTime = SystemClock.elapsedRealtime();
+            scheduleSeekBarUpdate();
+        } else {
+            stopSeekBarUpdate();
+        }
     }
 
     // ConnectionStateCallback Methods
@@ -268,6 +323,7 @@ public class MusicPlayer implements ConnectionStateCallback, PlayerNotificationC
                 playNextSong();
                 break;
             case LOST_PERMISSION:
+                stopSeekBarUpdate();
                 mIsPlaying = false;
                 mEventManager.postEvent(new LostPermissionEvent());
                 break;
@@ -277,5 +333,47 @@ public class MusicPlayer implements ConnectionStateCallback, PlayerNotificationC
     @Override
     public void onPlaybackError(ErrorType errorType, String s) {
 
+    }
+
+    // Time stuff
+    private void scheduleSeekBarUpdate() {
+        stopSeekBarUpdate();
+        if (!mExecutorService.isShutdown()) {
+            mScheduledFuture = mExecutorService.scheduleAtFixedRate(
+                    new Runnable() {
+                        @Override
+                        public void run() {
+                            mHandler.post(mUpdateProgressTask);
+                        }
+                    }, PROGRESS_UPDATE_INITIAL_INTERVAL,
+                    PROGRESS_UPDATE_INTERNAL, TimeUnit.MILLISECONDS
+            );
+        }
+    }
+
+    private void stopSeekBarUpdate() {
+        if (mScheduledFuture != null) {
+            mScheduledFuture.cancel(false);
+        }
+    }
+
+    private void updateProgress() {
+        if (!mIsPlaying) {
+            return;
+        }
+        long currentTime = SystemClock.elapsedRealtime();
+
+        long difference = currentTime - mPreviousTime;
+        mPreviousTime = currentTime;
+        mPreviousSongTime += difference;
+
+        mEventManager.postEvent(new UpdateSongTimeEvent(mPreviousSongTime));
+    }
+
+    @Subscribe
+    public void onDestroyPlayer(DestroyPlayerEvent event) {
+        Spotify.destroyPlayer(this);
+        stopSeekBarUpdate();
+        mExecutorService.shutdown();
     }
 }
