+++ app/src/main/java/me/mikecasper/musicvoice/services/musicplayer/events/DestroyPlayerEvent.java
@@ -1,0 +1,5 @@
+package me.mikecasper.musicvoice.services.musicplayer.events;
+
+public class DestroyPlayerEvent {
+
+}
