+++ app/src/main/java/me/mikecasper/musicvoice/services/musicplayer/events/StopSeekbarUpdateEvent.java
@@ -1,0 +1,4 @@
+package me.mikecasper.musicvoice.services.musicplayer.events;
+
+public class StopSeekbarUpdateEvent {
+}
